requirements.txt
aiohttp==0.19.0
RPi.GPIO==0.5.11
